+++ mozilla_repo_urls/parser.py
@@ -2,11 +2,14 @@ import giturlparse
 
 from mozilla_repo_urls.platforms import ADDITIONAL_PLATFORMS
 
-from .errors import InvalidRepoUrlError
+from .errors import InvalidRepoUrlError, UnsupportedPlatformError
 from .result import RepoUrlParsed
 
 for i, platform in enumerate(ADDITIONAL_PLATFORMS):
     giturlparse.platforms.PLATFORMS.insert(i, platform)
+
+
+_SUPPORTED_PLAFORMS = ("hgmo", "github")
 
 
 def parse(url_string):
@@ -17,5 +20,8 @@ def parse(url_string):
 
     if not parsed_url.valid:
         raise InvalidRepoUrlError(url_string)
+
+    if parsed_url.platform not in _SUPPORTED_PLAFORMS:
+        raise UnsupportedPlatformError(url_string, platform, _SUPPORTED_PLAFORMS)
 
     return parsed_url