+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `parse()` fails automatically if URL couldn't be parsed
+- `parse()` fails automatically if repo platform is neither `hg.mozilla.org` nor `github.com`
 
 
 ## [0.0.1] - 2022-09-05